requirements.txt
apache-airflow==2.7.1
pymysql
dbt-core
dbt-mysql
pandas>=2.0.0
sqlalchemy
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pymongo==4.6.1
beautifulsoup4==4.12.3
emoji==2.12.1
lxml==5.3.0
